machines: Unify lists of ifaces for both virtual networks and vNICs
